13-advanced-splitting/requirements.txt
# LangChain Core
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5

# Experimental features (for SemanticChunker)
langchain-experimental>=0.0.50

# Token splitting
tiktoken>=0.5.0

# Document Loaders
pypdf>=3.0.0  # For PDF loading
unstructured>=0.10.0  # For Markdown loading
unstructured[markdown]>=0.10.0  # Markdown support
markdown>=3.4.0  # Required dependency for unstructured markdown support

# For environment variable management
python-dotenv>=1.0.0
